Added a Fix for MacOS/XCodeCompiler misbehaviour
When mixing Read and Write operation the pointers misbehave and seeks
to the end of file and write the data at the end of file instead.

FIX: after reading, seeking to the intended position after reading to
write, Even if your current position is your intended position fixed
the problem. However calling tellg() or tellp() fixed the problem too!

The problem doesn’t occur on Windows using minGW compiler, reason for
such bug is yet to be investigated.